moved 'PublishReportMaker' from control file
