+++ client/ayon_core/tools/publisher/models/__init__.py
@@ -1,0 +1,6 @@
+from .publish import PublishReportMaker
+
+
+__all__ = (
+    "PublishReportMaker",
+)
